# backend/artifacts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Clinimetria",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/v2/protocol": {
      "post": {
        "tags": [
          "v2"
        ],
        "summary": "Run Protocol V2",
        "description": "Execute v2 analysis protocol with support for advanced methods.\n\nSupports mixed effects, clustered correlation, and all standard methods.",
        "operationId": "run_protocol_v2_api_v1_v2_protocol_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProtocolV2Request"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Run Protocol V2 Api V1 V2 Protocol Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/v2/ai/suggest-tests": {
      "post": {
        "tags": [
          "ai"
        ],
        "summary": "Ai Suggest Tests",
        "operationId": "ai_suggest_tests_api_v1_v2_ai_suggest_tests_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AISuggestTestsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Ai Suggest Tests Api V1 V2 Ai Suggest Tests Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/v2/ai/analyze-design": {
      "post": {
        "tags": [
          "ai"
        ],
        "summary": "Ai Analyze Design",
        "operationId": "ai_analyze_design_api_v1_v2_ai_analyze_design_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AIAnalyzeDesignRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Ai Analyze Design Api V1 V2 Ai Analyze Design Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AIAnalyzeDesignRequest": {
        "properties": {
          "dataset_id": {
            "type": "string",
            "title": "Dataset Id",
            "description": "Dataset identifier"
          },
          "text": {
            "type": "string",
            "title": "Text",
            "description": "Research design description"
          },
          "protocol": {
            "anyOf": [
              {
                "items": {
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Protocol",
            "description": "Current protocol for context"
          },
          "preferences": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Preferences",
            "description": "Global preferences"
          }
        },
        "type": "object",
        "required": [
          "dataset_id",
          "text"
        ],
        "title": "AIAnalyzeDesignRequest"
      },
      "AISuggestTestsRequest": {
        "properties": {
          "dataset_id": {
            "type": "string",
            "title": "Dataset Id",
            "description": "Dataset identifier"
          },
          "protocol": {
            "items": {
              "type": "object"
            },
            "type": "array",
            "title": "Protocol",
            "description": "Current protocol for context"
          }
        },
        "type": "object",
        "required": [
          "dataset_id",
          "protocol"
        ],
        "title": "AISuggestTestsRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ProtocolV2Request": {
        "properties": {
          "dataset_id": {
            "type": "string",
            "title": "Dataset Id",
            "description": "Dataset identifier"
          },
          "protocol": {
            "type": "object",
            "title": "Protocol",
            "description": "Analysis protocol configuration"
          },
          "alpha": {
            "type": "number",
            "maximum": 0.1,
            "minimum": 0.01,
            "title": "Alpha",
            "description": "Significance level",
            "default": 0.05
          }
        },
        "type": "object",
        "required": [
          "dataset_id",
          "protocol"
        ],
        "title": "ProtocolV2Request",
        "description": "Request model for v2 analysis protocol."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}